fix!: update saveBlob in order to handle the autosave functionality
When Cool saves the opened file automatically then the system now
overwrites the current version instead of create a new one. This fix
prevents the generation of multiple versions even if the user leaves
the documents openend without modifying it.

Updated the response of the POST wopi in order to return only the
LastModifiedTime instead of all the docs attributes.

Updated the Files SDK version: now it handles correctly the response of
the upload-version API.