fix: DOCS-167 Remove access token from the returned wopi endpoint (#9)
When docs-editor cleans up its environment from the last session in order
to reopen a document already opened before, the comunication with
carbonio-docs-connector fails sistematically due to an invalid token used
to call the API.

Now the access_token is returned only in the url used to redirect the openFile
request to the carbonio-docs-editor and it has been removed from the wopi
endpoint. When the carbonio-docs-editor cleans up the environment it makes
the request to carbonio-docs-connector passing the token formed correctly.