feat: DOCS-177 - Add logging system using logback
Now the system use logback to print on STDOUT and on FILE all the logs.
The rolling mechanism has been configured in order to have a daily log files.

In all codebase, the throwable of the raised exception has been added to the
logger.error statement in order to print the stacktrace.