refactor: DOCS-183 - Add OpenDocumentRepository interface
The OpenDocumentRespository interface is added to abstract how the
OpenDocumentToken is saved. In the CE project flavour, it is saved in the
already existing Cache. Now all the codebase use the implementation of
the OpenDocumentRepository instead of using the cache directly.

This will allow to have multiple implementations of how to save the
OpenDocumentToken.

ref: DOCS-183