refactor: add OpenDocumentTokenRepository interface (#15)
The OpenDocumentTokenRespository interface is added to abstract how the
OpenDocumentToken is saved. In the CE project flavour, it is saved in the
already existing Cache. Now all the codebase use the implementation of
the OpenDocumentTokenRepository instead of using the cache directly.

This will allow to have multiple implementations of how to save the
OpenDocumentToken.

refs: DOCS-183