feat: Add the requester locale during the docs-editor URL generation
This feature allows to generate the docs-editor URL specifing the
language of the menu. The localization is the one specified by the user
in the Carbonio settings and it is fetched along with all others user
attributes via a new API exposed by the UserManagement called /users/myself
This API returns all the attributes associated to the users that wants to open
the document.

- Added the Simulator to have the test environment to make the ITs
- Added the OpenFileApiIT to test the success of the /files/open API

refs: DOCS-191